requirements.txt
django==2.1.7
feedparser
requests
django-el-pagination
elasticsearch
django-elasticsearch-dsl
nltk
pandas
numpy
scikit-learn
social-auth-app-django
tqdm
djangoajax
django-allauth
psycopg2
django-autocomplete-light
django-angular
djangorestframework
markdown
django-filter
django-npm